Split interface into separate file
